+++ doc/release-notes.md
@@ -5,3 +5,9 @@ Bitcoin ABC version 0.22.4 is now available from:
   <https://download.bitcoinabc.org/0.22.4/>
 
 This release includes the following features and fixes:
+
+Command-line options
+--------------------
+
+-  The `-debug=db` logging category has been renamed to `-debug=walletdb`, to distinguish it from `coindb`.
+   `-debug=db` has been deprecated and will be removed in a next release.